v3.5 | Fixed an issue with PlayerInfoPacket getter in PacketUtils, implemented FireworkNPC, changed CrossbowUtils and more

+++ src/main/java/me/mohamad82/ruom/Testings.java
@@ -300,7 +300,7 @@ public final class Testings extends RUoMPlugin implements CommandExecutor {
         }
 
         @Override
-        protected void onStopUseItem(Player player, ItemStack item, float holdTime) {
+        protected void onStopUseItem(Player player, ItemStack item, int holdTime) {
             npc.stopUsingItem();
         }
